// src/shared/enums/bets-types.enum.ts
export enum BetsTypes {
  PLENO = 'plenoNumbers',
  SEMI_PLENO = 'semiPlenoNumbers',
  CALLE = 'calleNumbers',
  CUADRO = 'cuadroNumbers',
  LINEA = 'lineaNumbers',
  EVEN_ODD = 'even_odd',
  COLOR = 'color',
  COLUMN = 'columns',
  DOZEN = 'dozens',
  CHANCE_SIMPLE = 'chanceSimple',
  CUBRE = 'cubre',
  SPECIAL_CALLE = 'specialCalle',
}